Object Mapper class for POJO's/response from REST Service modified for new POJO's
